requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
idna==2.10
pycparser==2.20
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
Werkzeug==1.0.1
